merso/requirements.txt
weaviate-client>=3.26.7;<4.0.0
langchain
langchain-community
langchain-openai
python-dotenv